// frontend/Angular/src/app/generated/api/model/orderStatus.ts
/**
 * Pizza Ordering API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export const OrderStatus = {
    Confirmed: 'Confirmed',
    Making: 'Making',
    Complete: 'Complete',
    SentOutForDelivery: 'SentOutForDelivery'
} as const;
export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];